Begin restructure to divide client/server logic
Loosely based on https://github.com/Wavesonics/GodotClientServer.
Also brings the codebase closer to the recommended Godot practice
of storing files close to their use due to the elimination of the
dedicated res://src/scripts/ directory.

The move is incomplete and will be fleshed out further in commits
to come.